haze: split operations by type
